backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = []
license = "CC0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-ninja = "^1.1.0"
django-cors-headers = "^4.3.1"
django = "^5.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
